docker_compose_full_stack/web/dockering/requirements.txt
Django==1.11.15
djangorestframework==3.8.2
django-redis==4.9.0
redis==2.10.6
psycopg2==2.7.5
gunicorn==19.3.0
